Created the main data loader file and a config file
We can add local and training parameters in the config file.
For the data loader it is required to have the airbus-ship-detection.zip - that contains all the data and metadata - unzipped.
The data loader is a class called ShipDatasetLoader that has to be initialized with the training images directory and the csv containing the masks for each image.
We can use train_val_split() function to create the tf.Datasets for training and validation.
Everything is commented in the file.